extract server out to app file and remove automatic start
